Support Integer literal
Some systems might require CSV files to be set with exact literals.
This patch introduces the module, `Types::Literal`, and makes it provide
our own custom methods for built-in classes.

The scope of this patch is to support Integer literals, and the users of
this gem can declare specific literals for some columns like this:

```ruby
StrongCSV.new do
  let :foo, 123 # The column `:foo` must be set with `123`
end
```