Change Union syntax from `|` to `,`
I changed my mind that we should not require developers to use `using`
in their source code because it would not be intuitive for them.

`,` is just a method argument separator, so this gem does not seem to
affect users code base.

From this patch, this declaration means ":user_id must be 1, 2, or 3".

```ruby
let :user_id, 1, 2, 3
```