Use i18n to let users have own error messages
Error messages are important to let end-users know what happened during
the process of a CSV file. Some users might be non-English spearkers,
and it would be helpful to provide error messages in their native
language.

So, I added the dependency of i18n since i18n might be popular gem,
which is also used in Ruby on Rails.