Add Let#pick to declare value picker and support constraint in Integer
Users sometimes might want to pick data from a database with the
specified IDs. For example, when a user want to receive `:user_id`
and want to make sure the value really exists in the database.

In this case, the new rule `constraint` would be useful because it can
ensure the casted value satisfies the specified constraint. However,
accessing to the database per CSV row causes N+1. In order to avoid this
problem, `Let#pick` is also introduced in this patch. `Let#pick` is
intended for storing data in advance of evaluating each CSV row and lets
users put limitation to Integer validation with the stored data.

For now, constraint is only supported in `Types::Integer`, but I want to
give this feature to other types.