Support :error_message on #let
Users of strong_csv might want to show their custom error message
when the provided CSV is invalid. This patch introduces the new
parameter `:error_message` on `#let` in order to allow such a feature.

I changed the type of `Let#types` from `Hash` to `Array` and made each
element an instance of a new class `TypeWrapper`, which holds `type` as
well as other metadata, such as `:name` and `:error_message`. When I
first started to implement this feature, I wondered where should I put
the custom error message. Then, I came up with the idea of refactoring
`Let#types` and including metadata to each element of the collection. I
hope this works well.